Links now work in 'post_date asc' posts order
Made query of WordPress posts table more specific, looking for
post_type='post' and post_status='publish' when determining the
post_date of the next post in order past the ones displayed on the page.
This fixes a bug that was messing up the date range of recommended links
to load.

Made copy of WordPress WP_List_Table class into plugin.  This is
recommended in the codex for stability so that changes to WP_List_Table
in future WordPress releases don't break the plugin.